feature:[DO-5097] Implemented creating Alerts from the dashboard
