+++ frontend/src/app/modules/page/parts/parts.module.ts
@@ -32,12 +32,13 @@ import { TemplateModule } from '@shared/template.module';
 import { PartsRoutingModule } from './parts.routing';
 import { PartsComponent } from './presentation/parts.component';
 import { RelationComponent } from './presentation/relation/relation.component';
-import {AngularSplitModule} from "angular-split";
-import {BomLifecycleSettingsService} from "@shared/service/bom-lifecycle-settings.service";
+import { AngularSplitModule } from "angular-split";
+import { BomLifecycleSettingsService } from "@shared/service/bom-lifecycle-settings.service";
+import { OwnPartsComponent } from './presentation/own-parts/own-parts.component';
 
 @NgModule({
-  declarations: [PartsComponent, RelationComponent],
-    imports: [CommonModule, TemplateModule, SharedModule, PartsRoutingModule, RelationsModule, PartDetailsModule, AngularSplitModule],
+  declarations: [PartsComponent, RelationComponent, OwnPartsComponent],
+  imports: [CommonModule, TemplateModule, SharedModule, PartsRoutingModule, RelationsModule, PartDetailsModule, AngularSplitModule],
   providers: [PartsState, BomLifecycleSettingsService, PartsFacade, FormatPartSemanticDataModelToCamelCasePipe, ...getI18nPageProvider(['page.parts', 'partDetail'])],
 })
-export class PartsModule {}
+export class PartsModule { }
