+++ frontend/src/app/modules/page/parts/presentation/parts.component.ts
@@ -20,36 +20,21 @@
  ********************************************************************************/
 
 import { AfterViewInit, Component, OnDestroy, OnInit, QueryList, ViewChildren } from '@angular/core';
-import { Pagination } from '@core/model/pagination.model';
 import { PartsFacade } from '@page/parts/core/parts.facade';
 import { MainAspectType } from '@page/parts/model/mainAspectType.enum';
 import {
-  AssetAsBuiltFilter,
-  AssetAsDesignedFilter,
-  AssetAsPlannedFilter,
-  AssetAsRecycledFilter,
-  AssetAsSupportedFilter,
-  AssetAsOrderedFilter,
   Part,
 } from '@page/parts/model/parts.model';
-import { PartTableType, TableEventConfig, TableHeaderSort } from '@shared/components/table/table.model';
-import { View } from '@shared/model/view.model';
+import { PartTableType } from '@shared/components/table/table.model';
 import { PartDetailsFacade } from '@shared/modules/part-details/core/partDetails.facade';
 import { StaticIdService } from '@shared/service/staticId.service';
-import { BehaviorSubject, Observable, Subject } from 'rxjs';
+import { BehaviorSubject, Subject } from 'rxjs';
 import { BomLifecycleSize } from '@shared/components/bom-lifecycle-activator/bom-lifecycle-activator.model';
 import { BomLifecycleSettingsService, UserSettingView } from '@shared/service/bom-lifecycle-settings.service';
-import { toAssetFilter, toGlobalSearchAssetFilter } from '@shared/helper/filter-helper';
 import { FormControl, FormGroup } from '@angular/forms';
 import { ToastService } from '@shared/components/toasts/toast.service';
-import { PartsTableComponent } from '@shared/components/parts-table/parts-table.component';
-import { MatDialog } from '@angular/material/dialog';
-import { PARTS_BASE_ROUTE, getRoute } from '@core/known-route';
-import { Router } from '@angular/router';
 import { SearchHelper } from '@shared/helper/search-helper';
-import { DatePipe } from '@angular/common';
-import { RequestStepperComponent } from '@shared/components/request-notification/request-stepper/request-stepper.component';
-import { RequestContext } from '@shared/components/request-notification/request-notification.base';
+import { OwnPartsComponent } from './own-parts/own-parts.component';
 
 @Component({
   selector: 'app-parts',
@@ -59,32 +44,12 @@ import { RequestContext } from '@shared/components/request-notification/request-
 export class PartsComponent implements OnInit, OnDestroy, AfterViewInit {
   public readonly titleId = this.staticIdService.generateId('PartsComponent.title');
 
-  public readonly partsAsBuilt$: Observable<View<Pagination<Part>>>;
-  public readonly partsAsPlanned$: Observable<View<Pagination<Part>>>;
-  public readonly partsAsDesigned$: Observable<View<Pagination<Part>>>;
-  public readonly partsAsOrdered$: Observable<View<Pagination<Part>>>;
-  public readonly partsAsSupported$: Observable<View<Pagination<Part>>>;
-  public readonly partsAsRecycled$: Observable<View<Pagination<Part>>>;
-
   public readonly deselectPartTrigger$ = new Subject<Part[]>();
   public readonly addPartTrigger$ = new Subject<Part>();
   public readonly currentSelectedItems$ = new BehaviorSubject<Part[]>([]);
+
   public readonly searchListAsBuilt: string[];
   public readonly searchListAsPlanned: string[];
-
-  public tableAsBuiltSortList: TableHeaderSort[];
-  public tableAsPlannedSortList: TableHeaderSort[];
-  public tableAsDesignedSortList: TableHeaderSort[];
-  public tableAsSupportedSortList: TableHeaderSort[];
-  public tableAsOrderedSortList: TableHeaderSort[];
-  public tableAsRecycledSortList: TableHeaderSort[];
-
-  public assetAsBuiltFilter: AssetAsBuiltFilter;
-  public assetAsPlannedFilter: AssetAsPlannedFilter;
-  public assetAsDesignedFilter: AssetAsDesignedFilter;
-  public assetAsRecycledFilter: AssetAsRecycledFilter;
-  public assetAsSupportedFilter: AssetAsSupportedFilter;
-  public assetAsOrderedFilter: AssetAsOrderedFilter;
 
   public DEFAULT_PAGE_SIZE = 50;
   public ctrlKeyState = false;
@@ -95,480 +60,54 @@ export class PartsComponent implements OnInit, OnDestroy, AfterViewInit {
   protected readonly MainAspectType = MainAspectType;
   public readonly searchHelper = new SearchHelper();
 
-  @ViewChildren(PartsTableComponent) partsTableComponents: QueryList<PartsTableComponent>;
+  @ViewChildren(OwnPartsComponent) ownPartsComponents: QueryList<OwnPartsComponent>;
 
   constructor(
     private readonly partsFacade: PartsFacade,
     private readonly partDetailsFacade: PartDetailsFacade,
     private readonly staticIdService: StaticIdService,
-    private readonly userSettingService: BomLifecycleSettingsService,
-    public dialog: MatDialog,
-    private readonly router: Router,
+    public userSettings: BomLifecycleSettingsService,
     public toastService: ToastService,
-    public datePipe: DatePipe,
-  ) {
-    this.partsAsBuilt$ = this.partsFacade.partsAsBuilt$;
-    this.partsAsPlanned$ = this.partsFacade.partsAsPlanned$;
-    this.partsAsDesigned$ = this.partsFacade.partsAsDesigned$;
-    this.partsAsOrdered$ = this.partsFacade.partsAsOrdered$;
-    this.partsAsSupported$ = this.partsFacade.partsAsSupported$;
-    this.partsAsRecycled$ = this.partsFacade.partsAsRecycled$;
+  ) { }
 
-    this.tableAsBuiltSortList = [];
-    this.tableAsPlannedSortList = [];
-    this.tableAsDesignedSortList = [];
-    this.tableAsOrderedSortList = [];
-    this.tableAsSupportedSortList = [];
-    this.tableAsRecycledSortList = [];
-    this.searchListAsBuilt = [
-      'id',
-      'idShort',
-      'nameAtManufacturer',
-      'manufacturerName',
-      'manufacturerPartId',
-      'customerPartId',
-      'classification',
-      'nameAtCustomer',
-      'semanticDataModel',
-      'semanticModelId',
-      'manufacturingDate',
-      'manufacturingCountry',
-    ];
-    this.searchListAsPlanned = [
-      'id',
-      'idShort',
-      'nameAtManufacturer',
-      'manufacturerName',
-      'manufacturerPartId',
-      'classification',
-      'semanticDataModel',
-      'semanticModelId',
-      'validityPeriodFrom',
-      'validityPeriodTo',
-      'function',
-      'catenaXSiteId',
-      'functionValidFrom',
-      'functionValidUntil',
-    ];
-
-    this.assetAsBuiltFilter = {};
-    this.assetAsDesignedFilter = {};
-    this.assetAsOrderedFilter = {};
-    this.assetAsPlannedFilter = {};
-    this.assetAsRecycledFilter = {};
-    this.assetAsSupportedFilter = {};
-
-    window.addEventListener('keydown', event => {
-      this.ctrlKeyState = event.ctrlKey;
-    });
-    window.addEventListener('keyup', event => {
-      this.ctrlKeyState = event.ctrlKey;
-    });
-  }
-
-  public bomLifecycleSize: BomLifecycleSize = this.userSettingService.getSize(UserSettingView.PARTS);
+  public bomLifecycleSize: BomLifecycleSize = this.userSettings.getSize(UserSettingView.PARTS);
 
   public searchFormGroup = new FormGroup({});
   public searchControl: FormControl;
 
   public ngOnInit(): void {
-    this.partsFacade.setPartsAsBuilt();
-    this.partsFacade.setPartsAsPlanned();
     this.searchFormGroup.addControl('partSearch', new FormControl([]));
     this.searchControl = this.searchFormGroup.get('partSearch') as unknown as FormControl;
-  }
-
-  openDialog(): void {
-    const dialogRef = this.dialog.open(RequestStepperComponent, {
-      autoFocus: false,
-      data: {
-        selectedItems: this.currentSelectedItems$.value,
-        context: RequestContext.REQUEST_ALERT,
-        tabIndex: 1,
-        fromExternal: true,
-      },
-    });
-
-    const callback = (part: Part) => {
-      this.deselectPartTrigger$.next([part]);
-    };
-
-    dialogRef?.componentInstance.deselectPart.subscribe(callback);
-    if (dialogRef?.afterClosed) {
-      dialogRef.afterClosed().subscribe((_part: Part) => {
-        dialogRef.componentInstance.deselectPart.unsubscribe();
-      });
-    }
-  }
-
-  openDetailPage(part: Part): void {
-    const { link } = getRoute(PARTS_BASE_ROUTE);
-    this.router.navigate([`/${link}/${part.id}`], { queryParams: { type: part.mainAspectType } });
-  }
-
-  filterActivated(type: MainAspectType, assetFilter: any): void {
-    this.globalSearchActive = false;
-    switch (type) {
-      case MainAspectType.AS_BUILT: {
-        this.assetAsBuiltFilter = assetFilter;
-        this.partsFacade.setPartsAsBuilt(
-          0,
-          this.DEFAULT_PAGE_SIZE,
-          this.tableAsBuiltSortList,
-          toAssetFilter(this.assetAsBuiltFilter, true),
-          this.globalSearchActive,
-        );
-        break;
-      }
-      case MainAspectType.AS_PLANNED: {
-        this.assetAsPlannedFilter = assetFilter;
-        this.partsFacade.setPartsAsPlanned(
-          0,
-          this.DEFAULT_PAGE_SIZE,
-          this.tableAsPlannedSortList,
-          toAssetFilter(this.assetAsPlannedFilter, false),
-          this.globalSearchActive,
-        );
-        break;
-      }
-      case MainAspectType.AS_DESIGNED: {
-        this.assetAsDesignedFilter = assetFilter;
-        this.partsFacade.setPartsAsDesigned(
-          0,
-          this.DEFAULT_PAGE_SIZE,
-          this.tableAsDesignedSortList,
-          toAssetFilter(this.assetAsDesignedFilter, true),
-          this.globalSearchActive,
-        );
-        break;
-      }
-      case MainAspectType.AS_ORDERED: {
-        this.assetAsOrderedFilter = assetFilter;
-        this.partsFacade.setPartsAsOrdered(
-          0,
-          this.DEFAULT_PAGE_SIZE,
-          this.tableAsOrderedSortList,
-          toAssetFilter(this.assetAsOrderedFilter, true),
-          this.globalSearchActive,
-        );
-        break;
-      }
-      case MainAspectType.AS_SUPPORTED: {
-        this.assetAsSupportedFilter = assetFilter;
-        this.partsFacade.setPartsAsSupported(
-          0,
-          this.DEFAULT_PAGE_SIZE,
-          this.tableAsSupportedSortList,
-          toAssetFilter(this.assetAsSupportedFilter, true),
-          this.globalSearchActive,
-        );
-        break;
-      }
-      case MainAspectType.AS_RECYCLED: {
-        this.assetAsRecycledFilter = assetFilter;
-        this.partsFacade.setPartsAsRecycled(
-          0,
-          this.DEFAULT_PAGE_SIZE,
-          this.tableAsRecycledSortList,
-          toAssetFilter(this.assetAsRecycledFilter, true),
-          this.globalSearchActive,
-        );
-        break;
-      }
-    }
-  }
-
-  // TODO implement search for other tables when they are implemented
-  triggerPartSearch() {
-    this.searchHelper.resetFilterAndShowToast(true, this.partsTableComponents, this.toastService);
-    const searchValue = this.searchFormGroup.get('partSearch').value;
-    if (searchValue && searchValue !== '') {
-      this.globalSearchActive = true;
-      this.assetAsBuiltFilter = toGlobalSearchAssetFilter(searchValue, true, this.searchListAsBuilt, this.datePipe);
-      this.assetAsPlannedFilter = toGlobalSearchAssetFilter(searchValue, false, this.searchListAsPlanned, this.datePipe);
-      this.partsFacade.setPartsAsBuilt(
-        0,
-        this.DEFAULT_PAGE_SIZE,
-        this.tableAsPlannedSortList,
-        this.assetAsBuiltFilter,
-        this.globalSearchActive,
-      );
-      this.partsFacade.setPartsAsPlanned(
-        0,
-        this.DEFAULT_PAGE_SIZE,
-        this.tableAsBuiltSortList,
-        this.assetAsPlannedFilter,
-        this.globalSearchActive,
-      );
-    } else {
-      this.globalSearchActive = false;
-      this.assetAsBuiltFilter = {};
-      this.assetAsPlannedFilter = {};
-      this.partsFacade.setPartsAsBuilt(0, this.DEFAULT_PAGE_SIZE);
-      this.partsFacade.setPartsAsPlanned(0, this.DEFAULT_PAGE_SIZE);
-    }
-  }
-
-  public ngAfterViewInit(): void {
-    this.handleTableActivationEvent(this.bomLifecycleSize);
   }
 
   public ngOnDestroy(): void {
     this.partsFacade.unsubscribeParts();
   }
 
-  public onSelectItem($event: Record<string, unknown>, type: MainAspectType): void {
-    const selectedPart = $event as unknown as Part;
-    this.partDetailsFacade.mainAspectType = type;
-    this.partDetailsFacade.selectedPart = selectedPart;
-    this.openDetailPage(selectedPart);
-  }
+  triggerPartSearch() {
+    this.resetFilterAndShowToast();
 
-  public onAsBuiltTableConfigChange({ page, pageSize, sorting }: TableEventConfig): void {
-    this.setTableSortingList(sorting, MainAspectType.AS_BUILT);
+    const searchValue = this.searchFormGroup.get('partSearch').value;
 
-    let pageSizeValue = this.DEFAULT_PAGE_SIZE;
-    if (pageSize !== 0) {
-      pageSizeValue = pageSize;
-    }
-    this.partsFacade.setPartsAsBuilt(
-      page,
-      pageSizeValue,
-      this.tableAsBuiltSortList,
-      toAssetFilter(this.assetAsBuiltFilter, true),
-      this.globalSearchActive,
-    );
-  }
-
-  public onAsPlannedTableConfigChange({ page, pageSize, sorting }: TableEventConfig): void {
-    this.setTableSortingList(sorting, MainAspectType.AS_PLANNED);
-    let pageSizeValue = this.DEFAULT_PAGE_SIZE;
-    if (pageSize !== 0) {
-      pageSizeValue = pageSize;
-    }
-
-    this.partsFacade.setPartsAsPlanned(
-      page,
-      pageSizeValue,
-      this.tableAsPlannedSortList,
-      toAssetFilter(this.assetAsPlannedFilter, false),
-      this.globalSearchActive,
-    );
-  }
-
-  public onDefaultPaginationSizeChange(pageSize: number) {
-    this.DEFAULT_PAGE_SIZE = pageSize;
-  }
-
-  private setTableSortingList(sorting: TableHeaderSort, partTable: MainAspectType): void {
-    // if a sorting Columnlist exists but a column gets resetted:
-    if (
-      !sorting &&
-      (this.tableAsBuiltSortList ||
-        this.tableAsPlannedSortList ||
-        this.tableAsDesignedSortList ||
-        this.tableAsOrderedSortList ||
-        this.tableAsSupportedSortList ||
-        this.tableAsRecycledSortList)
-    ) {
-      this.resetTableSortingList(partTable);
-      return;
-    }
-
-    // if CTRL is pressed at to sortList
-    if (this.ctrlKeyState) {
-      const [columnName] = sorting;
-      let tableSortList: TableHeaderSort[] = [];
-      switch (partTable) {
-        case MainAspectType.AS_BUILT: {
-          tableSortList = this.tableAsBuiltSortList;
-          break;
-        }
-        case MainAspectType.AS_PLANNED: {
-          tableSortList = this.tableAsPlannedSortList;
-          break;
-        }
-        case MainAspectType.AS_DESIGNED: {
-          tableSortList = this.tableAsDesignedSortList;
-          break;
-        }
-        case MainAspectType.AS_ORDERED: {
-          tableSortList = this.tableAsOrderedSortList;
-          break;
-        }
-        case MainAspectType.AS_SUPPORTED: {
-          tableSortList = this.tableAsSupportedSortList;
-          break;
-        }
-        case MainAspectType.AS_RECYCLED: {
-          tableSortList = this.tableAsRecycledSortList;
-          break;
-        }
-      }
-
-      // Find the index of the existing entry with the same first item
-      const index = tableSortList.findIndex(([itemColumnName]) => itemColumnName === columnName);
-
-      if (index !== -1) {
-        // Replace the existing entry
-        tableSortList[index] = sorting;
-      } else {
-        // Add the new entry if it doesn't exist
-        tableSortList.push(sorting);
-      }
-
-      switch (partTable) {
-        case MainAspectType.AS_BUILT: {
-          this.tableAsBuiltSortList = tableSortList;
-          break;
-        }
-        case MainAspectType.AS_PLANNED: {
-          this.tableAsPlannedSortList = tableSortList;
-          break;
-        }
-        case MainAspectType.AS_DESIGNED: {
-          this.tableAsDesignedSortList = tableSortList;
-          break;
-        }
-        case MainAspectType.AS_ORDERED: {
-          this.tableAsOrderedSortList = tableSortList;
-          break;
-        }
-        case MainAspectType.AS_SUPPORTED: {
-          this.tableAsSupportedSortList = tableSortList;
-          break;
-        }
-        case MainAspectType.AS_RECYCLED: {
-          this.tableAsRecycledSortList = tableSortList;
-          break;
-        }
-      }
-    }
-    // If CTRL is not pressed just add a list with one entry
-    else {
-      switch (partTable) {
-        case MainAspectType.AS_BUILT: {
-          this.tableAsBuiltSortList = [sorting];
-          break;
-        }
-        case MainAspectType.AS_PLANNED: {
-          this.tableAsPlannedSortList = [sorting];
-          break;
-        }
-        case MainAspectType.AS_DESIGNED: {
-          this.tableAsDesignedSortList = [sorting];
-          break;
-        }
-        case MainAspectType.AS_ORDERED: {
-          this.tableAsOrderedSortList = [sorting];
-          break;
-        }
-        case MainAspectType.AS_SUPPORTED: {
-          this.tableAsSupportedSortList = [sorting];
-          break;
-        }
-        case MainAspectType.AS_RECYCLED: {
-          this.tableAsRecycledSortList = [sorting];
-          break;
-        }
-      }
+    for (const ownPartsComponent of this.ownPartsComponents) {
+      ownPartsComponent.updateOwnParts(searchValue);
     }
   }
 
-  private resetTableSortingList(partTable: MainAspectType): void {
-    switch (partTable) {
-      case MainAspectType.AS_BUILT: {
-        this.tableAsBuiltSortList = [];
-        break;
+  private resetFilterAndShowToast() {
+    const resetComponents = (components: QueryList<OwnPartsComponent>) => {
+      for (const component of components) {
+        const filterIsSet = this.searchHelper.resetFilterForAssetComponents(component.partsTableComponents, false);
+        if (filterIsSet) {
+          this.toastService.info('parts.input.global-search.toastInfo');
+        }
       }
-      case MainAspectType.AS_PLANNED: {
-        this.tableAsPlannedSortList = [];
-        break;
-      }
-      case MainAspectType.AS_DESIGNED: {
-        this.tableAsDesignedSortList = [];
-        break;
-      }
-      case MainAspectType.AS_ORDERED: {
-        this.tableAsOrderedSortList = [];
-        break;
-      }
-      case MainAspectType.AS_SUPPORTED: {
-        this.tableAsSupportedSortList = [];
-        break;
-      }
-      case MainAspectType.AS_RECYCLED: {
-        this.tableAsRecycledSortList = [];
-        break;
-      }
-    }
+    };
+    resetComponents(this.ownPartsComponents);
   }
 
-  public onAsDesignedTableConfigChange({ page, pageSize, sorting }: TableEventConfig): void {
-    this.setTableSortingList(sorting, MainAspectType.AS_DESIGNED);
-
-    let pageSizeValue = this.DEFAULT_PAGE_SIZE;
-    if (pageSize !== 0) {
-      pageSizeValue = pageSize;
-    }
-
-    this.partsFacade.setPartsAsDesigned(
-      page,
-      pageSizeValue,
-      this.tableAsDesignedSortList,
-      toAssetFilter(this.assetAsDesignedFilter, true),
-      this.globalSearchActive,
-    );
-  }
-
-  public onAsOrderedTableConfigChange({ page, pageSize, sorting }: TableEventConfig): void {
-    this.setTableSortingList(sorting, MainAspectType.AS_ORDERED);
-
-    let pageSizeValue = this.DEFAULT_PAGE_SIZE;
-    if (pageSize !== 0) {
-      pageSizeValue = pageSize;
-    }
-
-    this.partsFacade.setPartsAsOrdered(
-      page,
-      pageSizeValue,
-      this.tableAsOrderedSortList,
-      toAssetFilter(this.assetAsOrderedFilter, true),
-      this.globalSearchActive,
-    );
-  }
-
-  public onAsSupportedTableConfigChange({ page, pageSize, sorting }: TableEventConfig): void {
-    this.setTableSortingList(sorting, MainAspectType.AS_SUPPORTED);
-
-    let pageSizeValue = this.DEFAULT_PAGE_SIZE;
-    if (pageSize !== 0) {
-      pageSizeValue = pageSize;
-    }
-    this.partsFacade.setPartsAsSupported(
-      page,
-      pageSizeValue,
-      this.tableAsSupportedSortList,
-      toAssetFilter(this.assetAsSupportedFilter, true),
-      this.globalSearchActive,
-    );
-  }
-
-  public onAsRecycledTableConfigChange({ page, pageSize, sorting }: TableEventConfig): void {
-    this.setTableSortingList(sorting, MainAspectType.AS_RECYCLED);
-
-    let pageSizeValue = this.DEFAULT_PAGE_SIZE;
-    if (pageSize !== 0) {
-      pageSizeValue = pageSize;
-    }
-
-    this.partsFacade.setPartsAsRecycled(
-      page,
-      pageSizeValue,
-      this.tableAsRecycledSortList,
-      toAssetFilter(this.assetAsRecycledFilter, true),
-      this.globalSearchActive,
-    );
+  public ngAfterViewInit(): void {
+    this.handleTableActivationEvent(this.bomLifecycleSize);
   }
 
   public handleTableActivationEvent(bomLifecycleSize: BomLifecycleSize) {
