+++ frontend/src/app/modules/shared/components/table/table.component.scss
@@ -265,6 +265,7 @@ table {
   @apply font-semiBold;
   @apply text-black;
   font-size: 16px;
+  text-transform: uppercase;
 }
 
 .notification-table--header {
