+++ frontend/src/assets/locales/de/common.json
@@ -155,7 +155,7 @@
     "generic": "Die Eingabe ist ungültig."
   },
   "requestAlert": {
-    "headline": "Qualitätswarnung starten",
+    "headline": "Qualitätswarnung erstellen",
     "textAreaDescription": "Fügen sie eine Beschreibung hinzu.",
     "success_one": "Sie haben eine Qualitätswarnung für {{count}} Produkt erstellt.",
     "success_other": "Sie haben eine Qualitätswarnung für {{count}} Produkte erstellt."
@@ -243,6 +243,7 @@
   "commonAlert": {
     "viewAll": "Zu allen Warnungen gehen",
     "noAlerts": "Keine aktuellen Qualitätswarnungen",
+    "selectedPartsActions": "Warnung erstellen",
     "noAlertsDescription": "Falls es Qualitätswarnungen gibt, werden sie hier angezeigt.",
     "create": "Warnung definieren",
     "tabs": {
